requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
auth
passlib
python-jose
bcrypt==4.0.1
